ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
numpy==1.26.2
scikit-learn==1.3.2
python-multipart==0.0.6
python-docx==1.0.1
PyPDF2==3.0.1
transformers==4.35.2
torch==2.1.1
python-jose==3.3.0
python-dotenv==1.0.0
requests==2.31.0 